requirements.txt
absl-py==0.9.0
alabaster==0.7.12
asn1crypto==0.24.0
astor==0.8.1
astroid==1.6.5
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cloudpickle==1.2.2
cryptography==2.3.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.14
entrypoints==0.2.3
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imagesize==1.1.0
imutils==0.5.3
ipykernel==4.10.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
jedi==0.12.1
jeepney==0.4.1
Jinja2==2.10.3
joblib==0.14.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==5.2.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==13.2.1
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.6.0
numpy==1.18.0
numpydoc==0.9.1
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==19.2
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.4
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.11.2
psutil==5.4.7
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.5.2
pylint==1.9.2
pyOpenSSL==18.0.0
pyparsing==2.4.6
PySocks==1.6.8
python-dateutil==2.8.1
pytz==2019.3
pyzmq==17.1.2
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
scikit-learn==0.22
scipy==1.4.1
SecretStorage==3.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.3.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.1
spyder-kernels==0.2.6
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.4
tornado==5.1.1
traitlets==4.3.2
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.4.1
wrapt==1.11.2
